Refactoring library into smaller components for faster compilation
